requirements.txt
Flask==3.0.0
translatepy==2.3
python-docx==0.8.11
PyPDF2==3.0.1